Refactor plotting code into modules
