+++ src/covid_model_seiir_pipeline/pipeline/diagnostics/model/__init__.py
@@ -1,0 +1,12 @@
+from covid_model_seiir_pipeline.pipeline.diagnostics.model.plot_version import (
+    Location,
+    PlotVersion,
+    make_plot_versions,
+)
+from covid_model_seiir_pipeline.pipeline.diagnostics.model.plotter import (
+    COLOR_MAP,
+    make_grid_plot,
+)
+from covid_model_seiir_pipeline.pipeline.diagnostics.model.pdf_merger import (
+    merge_pdfs,
+)
